requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.16.0
Markdown==3.8.2
sqlparse==0.5.3
